processors/preprocess-python/requirements.txt
numpy==1.24.2
pandas==1.5.3
PyYAML==6.0.1
redis==5.0.1
tqdm==4.64.1
